creating pkg/auth package and migrating pkg/clients/authentication code that shouldn't be there
