+++ pkg/auth/errors.go
@@ -1,4 +1,4 @@
-package authentication
+package auth
 
 import (
 	"errors"
