+++ pkg/auth/middleware.go
@@ -1,4 +1,4 @@
-package authentication
+package auth
 
 import (
 	"context"
@@ -22,7 +22,7 @@ const (
 // Middleware defines the interface for authentication-related middleware functions used
 // to secure and validate requests.
 //
-//go:generate mockgen -destination=./mocks/middleware_mock.go -package=authentication_mocks github.com/alexgrauroca/practice-food-delivery-platform/pkg/clients/authentication Middleware
+//go:generate mockgen -destination=./mocks/middleware_mock.go -package=authentication_mocks github.com/alexgrauroca/practice-food-delivery-platform/pkg/auth Middleware
 type Middleware interface {
 	RequireCustomer() gin.HandlerFunc
 }
